Relocate hoc
